chore: added hashRouter
it was created a react custom hook to avoiding duplicated code
